add basic MindComponent
