+++ src/DuckMind/Ducia.Game/Ducia.Game.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>netstandard2.0</TargetFramework>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\..\..\..\Glint\src\Glint\Glint.csproj" />
+      <ProjectReference Include="..\..\..\..\XNez\Nez.Portable\XNez.csproj" />
+      <ProjectReference Include="..\Ducia\Ducia.csproj" />
+    </ItemGroup>
+
+</Project>
