+++ src/DuckMind/DuckMind.sln
@@ -6,6 +6,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MindDemos", "MindDemos\Mind
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ducia.Tests", "Ducia.Tests\Ducia.Tests.csproj", "{E2A30787-0AD3-47AF-96BD-64F892B90CA7}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ducia.Game", "Ducia.Game\Ducia.Game.csproj", "{B67F7FE6-2F7C-482E-8C5A-BF91A19E1414}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -24,5 +26,9 @@ Global
 		{E2A30787-0AD3-47AF-96BD-64F892B90CA7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E2A30787-0AD3-47AF-96BD-64F892B90CA7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E2A30787-0AD3-47AF-96BD-64F892B90CA7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B67F7FE6-2F7C-482E-8C5A-BF91A19E1414}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B67F7FE6-2F7C-482E-8C5A-BF91A19E1414}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B67F7FE6-2F7C-482E-8C5A-BF91A19E1414}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B67F7FE6-2F7C-482E-8C5A-BF91A19E1414}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
